Add remaining components to the new Stories format.
